a prototype for the new view-model-receiver system
