+++ lexos/models/base_model.py
@@ -1,22 +1,18 @@
-from flask import request
-
-
 class BaseModel:
     def __init__(self):
-        """This is the base model for all the models
+        """Model is a set of classes to do computation.
 
-        used to handle requests and other common stuff
+        Each model has 2 kinds of members:
+            - the options get from receivers (possibly fake inputted options).
+            - the result of higher level class (possibly fake inputted result).
+        Each model has 1 public method with name "generate"
+            - This is the computation result for lower class or view to use.
         """
-        if request.json:  # check request.json (ajax request) first
-            self._option = request.json
-        elif request.form:  # fall back to request.form (form request)
-            self._option = request.form
-        else:  # no data send to the back end
-            self._option = None
+        pass
 
-    @property
-    def front_end_data(self) -> dict:
-        return self._option
+    def generate(self):
+        """The virtual method to generate result of the model"""
+        raise NotImplementedError
 
     def data_exists_in_requests(self, data_key: str) -> bool:
         """check whether the data key exists in the data sent from front end
