+++ lexos/reciever/file_manager_receiver.py
@@ -1,0 +1,11 @@
+from lexos.reciever.base_receiver import BaseReceiver
+
+
+class FileManagerReceiver(BaseReceiver):
+    def __init__(self):
+        """So far there is no frontend option for file manager"""
+        super().__init__()
+
+    def options(self):
+        """So far there is no frontend option for file manager"""
+        pass
